fix #141: hook_exceptions causes failure in Python3
I'm not convinced that this exception hook is working as intended or
that it's necessary. It's not clear to me why stdout has to be in
unbuffered mode (but that can be done by setting PYTHONUNBUFFERED=1
anyway). traceback.format_exception will produce text with newlines, so
it will still span multiple lines. It *will* print exceptions to stdout
instead of stderr, but that could be done with the logging config. I
don't like it, but rather than mess with it I'm just going to catch the
error and not touch it.