Fix #153: hook_exceptions causes uwsgi to hang in Python 3
This isn't the first time it's cause problems, and mentioned in
c784c38d14, I do not think it is necessary. So I'm removing it. This was
causing uwsgi to hang in Python 3 when pypicloud was configured to use
dynamodb, and it prevented uwsgi from logging anything to the console in
Python 3. It was incredibly hard to track down it seems too magic and
too brittle to be worth it.