Fix #145: Add warnings and documentation when bucket name has a dot in it
If the bucket name has a dot in it and no region name is specified, the
generated url will take you to a 301 which will break pip. From my
tests, if you specify the bucket region, boto will generate the correct
url. The fix here is to just add documentation and warnings so people
can find the easy solution when they hit this edge case.