Fixing an issue with graceful reload
The reload logic was treating packages as the same if their names and
versions matched. This breaks down if a package has multiple builds for
the same name + version (e.g. wheel, tar.gz, windows, mac, linux, etc).
The solution is to compare the filenames to judge equality instead of
comparing name + version.